+++ samples/System.IO.Pipelines.Samples/SampleBase/RawHttpServerSampleBase.cs
@@ -65,7 +65,7 @@ namespace System.IO.Pipelines.Samples
 
                     // Writing directly to pooled buffers
                     var output = connection.Output;
-                    var formatter = new OutputFormatter<IPipeWriter>(output, SymbolTable.InvariantUtf8);
+                    var formatter = new OutputFormatter<PipeWriter>(output, SymbolTable.InvariantUtf8);
                     formatter.Append("HTTP/1.1 200 OK");
                     formatter.Append("\r\nContent-Length: 13");
                     formatter.Append("\r\nContent-Type: text/plain");
