+++ src/System.IO.Pipelines.Compression/CompressionPipelineExtensions.cs
@@ -10,82 +10,82 @@ namespace System.IO.Pipelines.Compression
 {
     public static class CompressionPipelineExtensions
     {
-        public static IPipeReader DeflateDecompress(this IPipeReader reader, PipeOptions options)
+        public static PipeReader DeflateDecompress(this PipeReader reader, PipeOptions options)
         {
             var inflater = new ReadableDeflateTransform(ZLibNative.Deflate_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = inflater.Execute(reader, pipe.Writer);
             return pipe.Reader;
         }
 
-        public static IPipeReader DeflateCompress(this IPipeReader reader, PipeOptions options, CompressionLevel compressionLevel)
+        public static PipeReader DeflateCompress(this PipeReader reader, PipeOptions options, CompressionLevel compressionLevel)
         {
             var deflater = new WritableDeflateTransform(compressionLevel, ZLibNative.Deflate_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = deflater.Execute(reader, pipe.Writer);
             return pipe.Reader;
         }
 
-        public static IPipeReader GZipDecompress(this IPipeReader reader, PipeOptions options)
+        public static PipeReader GZipDecompress(this PipeReader reader, PipeOptions options)
         {
             var inflater = new ReadableDeflateTransform(ZLibNative.GZip_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = inflater.Execute(reader, pipe.Writer);
             return pipe.Reader;
         }
 
-        public static IPipeWriter GZipCompress(this IPipeWriter writer, PipeOptions options, CompressionLevel compressionLevel)
+        public static PipeWriter GZipCompress(this PipeWriter writer, PipeOptions options, CompressionLevel compressionLevel)
         {
             var deflater = new WritableDeflateTransform(compressionLevel, ZLibNative.GZip_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = deflater.Execute(pipe.Reader, writer);
             return pipe.Writer;
         }
 
-        public static IPipeReader GZipCompress(this IPipeReader reader, PipeOptions options, CompressionLevel compressionLevel)
+        public static PipeReader GZipCompress(this PipeReader reader, PipeOptions options, CompressionLevel compressionLevel)
         {
             var deflater = new WritableDeflateTransform(compressionLevel, ZLibNative.GZip_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = deflater.Execute(reader, pipe.Writer);
             return pipe.Reader;
         }
 
-        public static IPipeReader CreateDeflateDecompressReader(PipeOptions options, IPipeReader reader)
+        public static PipeReader CreateDeflateDecompressReader(PipeOptions options, PipeReader reader)
         {
             var inflater = new ReadableDeflateTransform(ZLibNative.Deflate_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = inflater.Execute(reader, pipe.Writer);
             return pipe.Reader;
         }
 
-        public static IPipeReader CreateDeflateCompressReader(PipeOptions options, IPipeReader reader, CompressionLevel compressionLevel)
+        public static PipeReader CreateDeflateCompressReader(PipeOptions options, PipeReader reader, CompressionLevel compressionLevel)
         {
             var deflater = new WritableDeflateTransform(compressionLevel, ZLibNative.Deflate_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = deflater.Execute(reader, pipe.Writer);
             return pipe.Reader;
         }
 
-        public static IPipeReader CreateGZipDecompressReader(PipeOptions options, IPipeReader reader)
+        public static PipeReader CreateGZipDecompressReader(PipeOptions options, PipeReader reader)
         {
             var inflater = new ReadableDeflateTransform(ZLibNative.GZip_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = inflater.Execute(reader, pipe.Writer);
             return pipe.Reader;
         }
 
-        public static IPipeWriter CreateGZipCompressWriter(PipeOptions options, IPipeWriter writer, CompressionLevel compressionLevel)
+        public static PipeWriter CreateGZipCompressWriter(PipeOptions options, PipeWriter writer, CompressionLevel compressionLevel)
         {
             var deflater = new WritableDeflateTransform(compressionLevel, ZLibNative.GZip_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = deflater.Execute(pipe.Reader, writer);
             return pipe.Writer;
         }
 
-        public static IPipeReader CreateGZipCompressReader(PipeOptions options, IPipeReader reader, CompressionLevel compressionLevel)
+        public static PipeReader CreateGZipCompressReader(PipeOptions options, PipeReader reader, CompressionLevel compressionLevel)
         {
             var deflater = new WritableDeflateTransform(compressionLevel, ZLibNative.GZip_DefaultWindowBits);
-            var pipe = new Pipe(options);
+            var pipe = new ResetablePipe(options);
             var ignore = deflater.Execute(reader, pipe.Writer);
             return pipe.Reader;
         }
@@ -99,7 +99,7 @@ namespace System.IO.Pipelines.Compression
                 _deflater = new Deflater(compressionLevel, bits);
             }
 
-            public async Task Execute(IPipeReader reader, IPipeWriter writer)
+            public async Task Execute(PipeReader reader, PipeWriter writer)
             {
                 List<MemoryHandle> handles = new List<MemoryHandle>();
 
@@ -210,7 +210,7 @@ namespace System.IO.Pipelines.Compression
                 _inflater = new Inflater(bits);
             }
 
-            public async Task Execute(IPipeReader reader, IPipeWriter writer)
+            public async Task Execute(PipeReader reader, PipeWriter writer)
             {
                 List<MemoryHandle> handles = new List<MemoryHandle>();
