+++ src/System.IO.Pipelines.Extensions/PipelineReaderExtensions.cs
@@ -9,7 +9,7 @@ namespace System.IO.Pipelines
 {
     public static class PipelineReaderExtensions
     {
-        public static ValueTask<int> ReadAsync(this IPipeReader input, ArraySegment<byte> destination)
+        public static ValueTask<int> ReadAsync(this PipeReader input, ArraySegment<byte> destination)
         {
             while (true)
             {
@@ -44,7 +44,7 @@ namespace System.IO.Pipelines
             return new ValueTask<int>(input.ReadAsyncAwaited(destination));
         }
 
-        public static async Task CopyToAsync(this IPipeReader input, IPipeWriter output)
+        public static async Task CopyToAsync(this PipeReader input, PipeWriter output)
         {
             while (true)
             {
@@ -75,7 +75,7 @@ namespace System.IO.Pipelines
             }
         }
 
-        private static async Task<int> ReadAsyncAwaited(this IPipeReader input, ArraySegment<byte> destination)
+        private static async Task<int> ReadAsyncAwaited(this PipeReader input, ArraySegment<byte> destination)
         {
             while (true)
             {
@@ -99,12 +99,12 @@ namespace System.IO.Pipelines
             }
         }
 
-        public static Task CopyToAsync(this IPipeReader input, Stream stream)
+        public static Task CopyToAsync(this PipeReader input, Stream stream)
         {
             return input.CopyToAsync(stream, 4096, CancellationToken.None);
         }
 
-        public static async Task CopyToAsync(this IPipeReader input, Stream stream, int bufferSize, CancellationToken cancellationToken)
+        public static async Task CopyToAsync(this PipeReader input, Stream stream, int bufferSize, CancellationToken cancellationToken)
         {
             // TODO: Use bufferSize argument
             while (!cancellationToken.IsCancellationRequested)
