+++ src/System.IO.Pipelines.Text.Primitives/PipelineWriterExtensions.cs
@@ -7,7 +7,7 @@ namespace System.IO.Pipelines.Text.Primitives
 {
     public static class PipelineWriterExtensions
     {
-        public static PipelineTextOutput AsTextOutput(this IPipeWriter writer, SymbolTable symbolTable)
+        public static PipelineTextOutput AsTextOutput(this PipeWriter writer, SymbolTable symbolTable)
         {
             return new PipelineTextOutput(writer, symbolTable);
         }
