--- src/System.IO.Pipelines/System/IO/Pipelines/IPipe.cs
@@ -1,12 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace System.IO.Pipelines
-{
-    public abstract class IPipe
-    {
-        public abstract IPipeReader Reader { get; }
-        public abstract IPipeWriter Writer { get; }
-    }
-}
