+++ tests/System.IO.Pipelines.Tests/PipeCompletionCallbacksTests.cs
@@ -26,7 +26,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var scheduler = new TestScheduler();
-            var pipe = new Pipe(new PipeOptions(_pool, writerScheduler: scheduler));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, writerScheduler: scheduler));
             pipe.Reader.Complete();
 
             pipe.Writer.OnReaderCompleted((exception, state) =>
@@ -44,7 +44,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var scheduler = new TestScheduler();
-            var pipe = new Pipe(new PipeOptions(_pool, readerScheduler: scheduler));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, readerScheduler: scheduler));
             pipe.Writer.Complete();
 
             pipe.Reader.OnWriterCompleted((exception, state) =>
@@ -60,7 +60,7 @@ namespace System.IO.Pipelines.Tests
         [Fact]
         public void OnReaderCompletedThrowsWithNullCallback()
         {
-            var pipe = new Pipe(new PipeOptions(_pool));
+            var pipe = new ResetablePipe(new PipeOptions(_pool));
 
             Assert.Throws<ArgumentNullException>(() => pipe.Writer.OnReaderCompleted(null, null));
         }
@@ -68,7 +68,7 @@ namespace System.IO.Pipelines.Tests
         [Fact]
         public void OnWriterCompletedThrowsWithNullCallback()
         {
-            var pipe = new Pipe(new PipeOptions(_pool));
+            var pipe = new ResetablePipe(new PipeOptions(_pool));
 
             Assert.Throws<ArgumentNullException>(() => pipe.Reader.OnWriterCompleted(null, null));
         }
@@ -78,7 +78,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var scheduler = new TestScheduler();
-            var pipe = new Pipe(new PipeOptions(_pool, writerScheduler: scheduler));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, writerScheduler: scheduler));
             pipe.Writer.OnReaderCompleted((exception, state) =>
             {
                 callbackRan = true;
@@ -94,7 +94,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var scheduler = new TestScheduler();
-            var pipe = new Pipe(new PipeOptions(_pool, readerScheduler: scheduler));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, readerScheduler: scheduler));
             pipe.Reader.OnWriterCompleted((exception, state) =>
             {
                 callbackRan = true;
@@ -110,7 +110,7 @@ namespace System.IO.Pipelines.Tests
         {
             var exception = new Exception();
             var scheduler = new TestScheduler();
-            var pipe = new Pipe(new PipeOptions(_pool, writerScheduler: scheduler));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, writerScheduler: scheduler));
             pipe.Writer.OnReaderCompleted((e, state) => throw exception, null);
             pipe.Reader.Complete();
 
@@ -124,7 +124,7 @@ namespace System.IO.Pipelines.Tests
         {
             var exception = new Exception();
             var scheduler = new TestScheduler();
-            var pipe = new Pipe(new PipeOptions(_pool, readerScheduler: scheduler));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, readerScheduler: scheduler));
             pipe.Reader.OnWriterCompleted((e, state) => throw exception, null);
             pipe.Writer.Complete();
 
@@ -138,7 +138,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var scheduler = new TestScheduler();
-            var pipe = new Pipe(new PipeOptions(_pool, writerScheduler: scheduler));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, writerScheduler: scheduler));
             pipe.Writer.OnReaderCompleted((exception, state) =>
             {
                 callbackRan = true;
@@ -163,7 +163,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var scheduler = new TestScheduler();
-            var pipe = new Pipe(new PipeOptions(_pool, readerScheduler: scheduler));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, readerScheduler: scheduler));
             pipe.Reader.OnWriterCompleted((exception, state) =>
             {
                 callbackRan = true;
@@ -187,7 +187,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var callbackState = new object();
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
             pipe.Writer.OnReaderCompleted((exception, state) =>
             {
                 Assert.Equal(callbackState, state);
@@ -203,7 +203,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var callbackState = new object();
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
             pipe.Reader.OnWriterCompleted((exception, state) =>
             {
                 Assert.Equal(callbackState, state);
@@ -222,7 +222,7 @@ namespace System.IO.Pipelines.Tests
             var callbackState3 = new object();
             var counter = 0;
 
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
             pipe.Writer.OnReaderCompleted((exception, state) =>
             {
                 Assert.Equal(callbackState1, state);
@@ -257,7 +257,7 @@ namespace System.IO.Pipelines.Tests
             var callbackState3 = new object();
             var counter = 0;
 
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
             pipe.Reader.OnWriterCompleted((exception, state) =>
             {
                 Assert.Equal(callbackState1, state);
@@ -294,7 +294,7 @@ namespace System.IO.Pipelines.Tests
 
             var counter = 0;
 
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
             pipe.Writer.OnReaderCompleted((exception, state) =>
             {
                 Assert.Equal(callbackState1, state);
@@ -327,7 +327,7 @@ namespace System.IO.Pipelines.Tests
 
             var counter = 0;
 
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
             pipe.Reader.OnWriterCompleted((exception, state) =>
             {
                 Assert.Equal(callbackState1, state);
@@ -354,7 +354,7 @@ namespace System.IO.Pipelines.Tests
         public void OnWriterCompletedPassesException()
         {
             var callbackRan = false;
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
             var readerException = new Exception();
 
             pipe.Reader.OnWriterCompleted((exception, state) =>
@@ -371,7 +371,7 @@ namespace System.IO.Pipelines.Tests
         public void OnReaderCompletedPassesException()
         {
             var callbackRan = false;
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
             var readerException = new Exception();
 
             pipe.Writer.OnReaderCompleted((exception, state) =>
@@ -389,7 +389,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var continuationRan = false;
-            var pipe = new Pipe(new PipeOptions(_pool) );
+            var pipe = new ResetablePipe(new PipeOptions(_pool) );
 
             pipe.Reader.OnWriterCompleted((exception, state) =>
             {
@@ -413,7 +413,7 @@ namespace System.IO.Pipelines.Tests
         {
             var callbackRan = false;
             var continuationRan = false;
-            var pipe = new Pipe(new PipeOptions(_pool, maximumSizeHigh: 5));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, maximumSizeHigh: 5));
 
             pipe.Writer.OnReaderCompleted((exception, state) =>
             {
@@ -439,7 +439,7 @@ namespace System.IO.Pipelines.Tests
         public void CompletingReaderFromWriterCallbackWorks()
         {
             var callbackRan = false;
-            var pipe = new Pipe(new PipeOptions(_pool, maximumSizeHigh: 5));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, maximumSizeHigh: 5));
 
             pipe.Writer.OnReaderCompleted((exception, state) =>
             {
@@ -459,7 +459,7 @@ namespace System.IO.Pipelines.Tests
         public void CompletingWriterFromReaderCallbackWorks()
         {
             var callbackRan = false;
-            var pipe = new Pipe(new PipeOptions(_pool, maximumSizeHigh: 5));
+            var pipe = new ResetablePipe(new PipeOptions(_pool, maximumSizeHigh: 5));
 
             pipe.Reader.OnWriterCompleted((exception, state) =>
             {
