+++ tests/System.IO.Pipelines.Tests/PipeLengthTests.cs
@@ -11,12 +11,12 @@ namespace System.IO.Pipelines.Tests
     public class PipeLengthTests : IDisposable
     {
         private MemoryPool _pool;
-        private Pipe _pipe;
+        private ResetablePipe _pipe;
 
         public PipeLengthTests()
         {
             _pool = new MemoryPool();
-            _pipe = new Pipe(new PipeOptions(_pool));
+            _pipe = new ResetablePipe(new PipeOptions(_pool));
         }
 
         public void Dispose()
