+++ tests/System.IO.Pipelines.Tests/PipeResetTests.cs
@@ -10,12 +10,12 @@ namespace System.IO.Pipelines.Tests
     public class PipeResetTests : IDisposable
     {
         private MemoryPool _pool;
-        private Pipe _pipe;
+        private ResetablePipe _pipe;
 
         public PipeResetTests()
         {
             _pool = new MemoryPool();
-            _pipe = new Pipe(new PipeOptions(_pool));
+            _pipe = new ResetablePipe(new PipeOptions(_pool));
         }
 
         public void Dispose()
